Fix triple faults in interrupt routine handlers
Previously, the interrupt routine handlers would incorrectly use the GDT
descriptor for code segments as opposed to data segments and result in a
triple fault as soon as interrupts are enabled on real hardware, as well
as some accurate emulators such as Bochs.

+++ kernel/include/arch/i386/descriptors/idt.h
@@ -19,5 +19,5 @@ struct idt_gate_debug {
 };
 
 void idt_install();
-void _idt_set_gate(uint8_t num, uint64_t base, uint16_t sel,
+void _idt_set_gate(uint8_t num, uint32_t base, uint16_t sel,
                    uint8_t flags);
